# Use vmware-rpctool instead of vmtoolsd for performance reasons