Fixes #18

`vmware-rpctool` doesn't hang when called repeatedly and is generally faster than `vmtoolsd`.
According to https://github.com/vmware/open-vm-tools/issues/377 it should be save to replace it.

Alternatively a fallback could be implemented in case `vmware-rpctool` is not available but all distros should include it as well.

I have not tested this in our environment yet, will do that tomorrow.