@yvespp 2019-10-30
I tested against my branch and everything works, cloud-init is now 4-6 seconds faster:
Before (`vmtoolsd`):
```
cloud-init analyze show
-- Boot Record 01 --
The total time elapsed since completing an event is printed after the "@" character.
The time the event takes is printed after the "+" character.

Starting stage: init-local
|`->no cache found @00.00600s +00.00000s
|`->found local data from DataSourceVMwareGuestInfo @00.04000s +06.12200s
Finished stage: (init-local) 06.36500 seconds

-- Boot Record 02 --
The total time elapsed since completing an event is printed after the "@" character.
The time the event takes is printed after the "+" character.

Starting stage: init-local
|`->cache invalid in datasource: DataSourceVMwareGuestInfo @00.00400s +00.03000s
|`->found local data from DataSourceVMwareGuestInfo @00.03600s +04.09300s
Finished stage: (init-local) 04.19000 seconds
```

After (`vmware-rpctool`):
```
root@yp-test-master-cb12345c442a:/# cloud-init analyze show
-- Boot Record 01 --
The total time elapsed since completing an event is printed after the "@" character.
The time the event takes is printed after the "+" character.

Starting stage: init-local
|`->no cache found @00.00700s +00.00100s
|`->found local data from DataSourceVMwareGuestInfo @00.07300s +00.02200s
Finished stage: (init-local) 00.33600 seconds
```

And total time:
```
root@yp-test-master-cb12345c442a:/# systemd-analyze blame
          1.633s cloud-init-local.service
          1.420s keyboard-setup.service
          1.401s cloud-init.service
          1.241s cloud-config.service
          1.080s docker.service
           797ms cloud-final.service
           640ms dev-sda1.device
```