@akutz 2019-11-11
Hi @yvespp,

Thank you for this PR! However, unless I'm mistaken, isn't `rpctool` a non-standard program that has to be built into the image? The command `vmtoolsd` is part of OpenVM tools, and thus is going to be in any image that has the tools installed.

I'd rather see a fail-gracefully set up where `rpctool` is used, if present, otherwise it fails gracefully back to using `vmtoolsd`.

Thoughts?