@yvespp 2019-11-12
@akutz 
> Thank you for this PR! However, unless I'm mistaken, isn't `rpctool` a non-standard program that has to be built into the image? The command `vmtoolsd` is part of OpenVM tools, and thus is going to be in any image that has the tools installed.

I asked in my issue in open-vm-tools if `rpctool` is also included in the distro packages of open-vm-tools and they said that it is. So it should be safe to use, see the two comments here: https://github.com/vmware/open-vm-tools/issues/377#issuecomment-543409940

But it should not be hard to implement a fallback so I will do that.