@akutz 2019-11-12
Hi @yvespp,

I didn't realize `rpctool` was included with OpenVM tools. If it is, then I think there should be nothing wrong with using it. Let me run some tests -- you don't need to amend this PR just yet, not unless you want to do so. Thanks!