@akutz 2019-11-12
Hi @dougm,

Can you please review this PR as well as confirm that the `rpctool` program should be available anywhere `vmtoolsd` is present? Thank you!